完成基本的后端接口

+++ model/baseModel.go
@@ -7,7 +7,7 @@ import (
 )
 
 type BaseModel struct {
-	ID         int32          `gorm:"primaryKey" json:"id"`
+	ID         int64          `gorm:"primaryKey" json:"id"`
 	CreateTime time.Time      `gorm:"column:create_time" json:"-"`
 	UpdateTime time.Time      `gorm:"column:update_time" json:"-"`
 	DeleteTime gorm.DeletedAt `gorm:"column:delete_time" json:"-"`
